third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-bottom-up-large-tree-search-expected.txt

Tests that search works for large bottom-up view of CPU profile.

10012: foo12 10 9890
10120: foo120 10 8810
10121: foo121 10 8800
10122: foo122 10 8790
10123: foo123 10 8780
10124: foo124 10 8770
10125: foo125 10 8760
10126: foo126 10 8750
10127: foo127 10 8740
10128: foo128 10 8730
10129: foo129 10 8720

Profiler was disabled.
